requirements.txt
# Core AWS and Bedrock
bedrock-agentcore>=1.1.0
boto3>=1.34.0

# MCP and Agents
mcp>=1.0.0
strands-agents>=1.0.0
strands-agents-tools

# Utilities
python-dotenv>=1.0.0

# Python package manager (provides uvx command for running MCP servers)
uv>=0.9.0